backend/requirements.txt
django
djangorestframework
environs
python-decouple
djangorestframework-simplejwt